Added options for space and modify callback
There is possibility to change the space character - i.e that which will
replace spaces. It's dash by default.

The second change is optional options parameter modify. It specifies
callback which will be fired before processing the text. It's useful
when you need to change the text in some way - add prefix, suffix etc.